# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /academic-calendar:
    get:
      tags: [Academic Calendar]
      operationId: listAcademicCalendar
      summary: List published weekly calendars
      responses:
        "200":
          description: Published academic calendars.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcademicCalendar"
    post:
      tags: [Academic Calendar]
      operationId: createAcademicCalendar
      summary: Publish a weekly academic calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcademicCalendarInput"
      responses:
        "201":
          description: Calendar published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcademicCalendar"
        "401":
          description: Authentication required.
  /academic-calendar/{id}:
    delete:
      tags: [Academic Calendar]
      operationId: deleteAcademicCalendar
      summary: Remove a weekly academic calendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Calendar removed.
        "401":
          description: Authentication required.
        "404":
          description: Calendar not found.
  /admission-applications:
    post:
      tags: [Admissions]
      operationId: createAdmissionApplication
      summary: Submit an admission application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionApplicationInput"
      responses:
        "201":
          description: Application received.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplicationReceipt"
        "400":
          description: Invalid application details.
    get:
      tags: [Admissions]
      operationId: listAdmissionApplications
      summary: List admission applications for school staff
      responses:
        "200":
          description: Admission applications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdmissionApplication"
        "401":
          description: Authentication required.
  /admission-applications/{id}:
    patch:
      tags: [Admissions]
      operationId: updateAdmissionApplication
      summary: Update an admission application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionApplicationStatusUpdate"
      responses:
        "200":
          description: Application updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplication"
        "400":
          description: Invalid application details.
        "401":
          description: Authentication required.
        "404":
          description: Application not found.
...
components:
...
  schemas:
...
    AcademicCalendar:
      type: object
      required: [id, title, weekStart, weekEnd, term, fileName, objectPath, contentType, fileSize, uploadedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        term:
          type: string
        fileName:
          type: string
        objectPath:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        uploadedAt:
          type: string
          format: date-time
    AcademicCalendarInput:
      type: object
      required: [title, weekStart, weekEnd, term, fileName, objectPath, contentType, fileSize]
      properties:
        title:
          type: string
          minLength: 1
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        term:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        contentType:
          type: string
          minLength: 1
        fileSize:
          type: integer
          minimum: 1
...
    AdmissionApplication:
      type: object
      required: [id, studentFirstName, studentLastName, grade, entryYear, guardianEmail, guardianPhone, status, submittedAt]
      properties:
        id:
          type: integer
        studentFirstName:
          type: string
        studentLastName:
          type: string
        grade:
          type: string
        entryYear:
          type: string
        guardianEmail:
          type: string
        guardianPhone:
          type: string
        status:
          type: string
        submittedAt:
          type: string
          format: date-time
    AdmissionApplicationInput:
      type: object
      required: [studentFirstName, studentLastName, grade, entryYear, guardianEmail, guardianPhone]
      properties:
        studentFirstName:
          type: string
          minLength: 1
        studentLastName:
          type: string
          minLength: 1
        grade:
          type: string
          minLength: 1
        entryYear:
          type: string
          minLength: 1
        guardianEmail:
          type: string
          minLength: 3
        guardianPhone:
          type: string
          minLength: 7
    AdmissionApplicationReceipt:
      type: object
      required: [id, studentFirstName, status, submittedAt]
      properties:
        id:
          type: integer
        studentFirstName:
          type: string
        status:
          type: string
        submittedAt:
          type: string
          format: date-time
    AdmissionApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          minLength: 1
...