# lib/api-zod/src/generated/types/admissionApplicationReceipt.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdmissionApplicationReceipt {
  id: number;
  studentFirstName: string;
  status: string;
  submittedAt: Date;
}